requirements.txt
fastapi==0.111.0 
uvicorn==0.29.0 
paddleocr==3.0.3 
paddlepaddle==3.0.0 
pdf2image==1.17.0 
pillow==10.3.0 
requests==2.31.0 
python-multipart==0.0.9